Extract useDismissable hook for menu dismissal
