+++ frontend/src/hooks/useDismissable.ts
@@ -1,0 +1,27 @@
+import { useEffect, useRef, useState } from 'react'
+
+// Open/close state for a popover-like element that dismisses on an outside
+// click or the Escape key. Attach the returned ref to the element whose bounds
+// define "inside".
+export function useDismissable<T extends HTMLElement = HTMLElement>() {
+  const [open, setOpen] = useState(false)
+  const ref = useRef<T>(null)
+
+  useEffect(() => {
+    if (!open) return
+    const onPointerDown = (e: PointerEvent) => {
+      if (ref.current && !ref.current.contains(e.target as Node)) setOpen(false)
+    }
+    const onKey = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') setOpen(false)
+    }
+    document.addEventListener('pointerdown', onPointerDown)
+    document.addEventListener('keydown', onKey)
+    return () => {
+      document.removeEventListener('pointerdown', onPointerDown)
+      document.removeEventListener('keydown', onKey)
+    }
+  }, [open])
+
+  return { open, setOpen, ref }
+}
